% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aamisc-package.R
\docType{package}
\name{aamisc-package}
\alias{aamisc}
\alias{aamisc-package}
\title{aamisc: Miscellaneous R Functions by Ali Altintas}
\description{
A collection of utility functions
}
\author{
Ali Altintas \email{altintas.ali@gmail.com}
}
\keyword{internal}
