% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbrev.R
\name{abbrev}
\alias{abbrev}
\title{Abbreviate character string using a maximum character limit}
\usage{
abbrev(x, n = 50)
}
\arguments{
\item{x}{A vector of character string.}

\item{n}{Maximum Number of characters to be displayed before suffixed with triple dots ("...")}
}
\value{
An abbreviated character string
}
\description{
Abbreviate character string using a maximum character limit
}
\examples{

a <- paste(letters, collapse = "")
abbrev(a, n = 0)
#[1] "a..."

abbrev(a, n = 4)
#[1] "abcd..."

}
