% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basenameLevel.R
\name{basenameLevel}
\alias{basenameLevel}
\title{Get directory root}
\usage{
basenameLevel(x, dirLevel = 0)
}
\arguments{
\item{x}{Path of a file or directory. Acts similar to \code{\link{basename}}}

\item{dirLevel}{Pointer to base directory. 0 means the base directory. 1 or bigger integer is upper directory in that level. See examples.}
}
\value{
Directory with the requested level information
}
\description{
Get directory root
}
\examples{
file <- "/home/projects/read.r"
basenameLevel(file)
# [1] "read.r"

basenameLevel(file, dirLevel = 1)
# [1] "projects/read.r"

#' basenameLevel(file, dirLevel = 2)
# [1] "home/projects/read.r"
}
