% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNAdf.R
\name{createNAdf}
\alias{createNAdf}
\title{Create an 1-row NA table from data.frame while keeping the column names. Works for empty data.frames as well.}
\usage{
createNAdf(df, onlyEmptyTable = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{onlyEmptyTable}{logical. If TRUE (default), only empty tables will be converted to 1-row NA table. Else, full table will be converted to 1-row NA table.}
}
\value{
1-row NA table.
}
\description{
Create an 1-row NA table from data.frame while keeping the column names. Works for empty data.frames as well.
}
\examples{

x <- matrix(seq(1:12), nrow = 3)
colnames(x) <- letters[1:4]
x <- data.frame(x)
print(x)

\dontrun{
## Throws an error as the data.frame is not empty.
createNAdf(df = x, onlyEmptyTable = TRUE)
}

## Setting the onlyEmptyTable to FALSE will ignore it
createNAdf(df = x, onlyEmptyTable = FALSE)
#   a  b  c  d
#1 NA NA NA NA

## Originally intended to use for empty data.frames
x2 <- x[0,]
createNAdf(df = x2, onlyEmptyTable = TRUE)
#   a  b  c  d
#1 NA NA NA NA
}
