% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_organism_by_ensemble_id.R
\name{detect_organism_by_ensemble_id}
\alias{detect_organism_by_ensemble_id}
\title{Get organism information from ENSEMBL ID}
\usage{
detect_organism_by_ensemble_id(ensembl_gene_id)
}
\arguments{
\item{ensembl_gene_id}{Ensemble gene, transcript or protein ID as character (e.g. ENSMUSG00000024411). Currently supports only human, mouse and rat}
}
\value{
List object (named \code{org}) contanining
\itemize{
   \item \code{$org} organism name (e.g. mouse)
   \item \code{$id} \href{http://www.genome.jp/kegg/catalog/org_list.html}{KEGG ID}  of the organism (e.g. mmu for mouse)
   \item \code{$db} org.db library in R (e.g. org.Mm.eg.db for mouse)
   \item \code{$scientific} Scientific name of the organism (e.g. 'Mus musculus' for mouse)
}
}
\description{
Get organism information from ENSEMBL ID
}
\examples{
detect_organism_by_ensemble_id(ensembl_gene_id = "ENSMUSG00000024411")

#$org
#[1] "mouse"
#
#$id
#[1] "mmu"
#
#$db
#[1] "org.Mm.eg.db"
#
#$scientific
#[1] "Mus musculus"
#
#$type
#[1] "gene"

}
