% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_col2rowname.R
\name{df_col2rowname}
\alias{df_col2rowname}
\title{Convert \code{data.frame} to \code{matrix} while moving the non-numeric values into \code{rownames}}
\usage{
df_col2rowname(df, col2rowname = NULL, sep = "_")
}
\arguments{
\item{df}{A \code{data.frame}}

\item{col2rowname}{Vector for the columns to be moved to rownames. Can be numeric (column ids) or character (column names)}

\item{sep}{Seperator for rownames if more than 1 element supplied in \code{col2rowname}}
}
\value{
A matrix object.
}
\description{
Convert \code{data.frame} to \code{matrix} while moving the non-numeric values into \code{rownames}
}
\examples{
# Create sample data.frame
x <- data.frame(Group1 = letters[1:5], ids = 1:5, id.squared = (1:5)^2)
x
X <- df_col2rowname(x)
X

# Create sample data.frame with 2 character columns
x2 <- data.frame(x, Group2 = LETTERS[6:10])
x2

X2 <- df_col2rowname(x2, col2rowname = c(1,4), sep = "_")
X2
X2 <- df_col2rowname(x2, col2rowname = c("Group1","Group2"), sep = "_")
X2

# Rownames can be reverse ordered
X3 <- df_col2rowname(x2, col2rowname = c(4,1), sep = "_")
X3
X3 <- df_col2rowname(x2, col2rowname = c("Group2","Group1"), sep = "_")
X3

}
