% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplotEnrich.R
\name{dotplotEnrich}
\alias{dotplotEnrich}
\title{Dotplot representation for enrichment results}
\usage{
dotplotEnrich(
  dt,
  topn = 10,
  topn.pref = c("qval", "dot"),
  qcut = 0.05,
  nchar = 60,
  direction = "DEstatus",
  group = "Condition",
  dot = "GeneRatio",
  dot.cex = 1,
  qval = "p.adjust",
  term.id = "ID",
  term.name = "Description",
  term.reverse = FALSE,
  plot.by = c("direction", "group")[1],
  onepanel = FALSE,
  arrows = FALSE,
  arrow.col = "beige"
)
}
\arguments{
\item{dt}{data.table or data.frame with the enrichment results. See details for necessary information.}

\item{topn}{Top \code{n} enriched terms to be plotted}

\item{topn.pref}{The ranking method to identify \code{topn} terms. Can be either \code{qval} or \code{dot}. \code{qval} sorts the dots by ascending order (lowest q-value first) and \code{dot} sorts the dots by descending order (e.g. biggest NES/GeneRatio) by absolute values.}

\item{qcut}{The adjusted p-value cutoff for enriched terms.}

\item{nchar}{Maximum characters to be shown on on y-axis. If the term contains more characters than \code{nchar}, it will be abbreviated with "(...)"}

\item{direction}{The column name in \code{dt} defining the direction  (e.g. up, down, all) of the enrichment}

\item{group}{The column name in \code{dt} defining the groups for each enrichment. The contents will be written on x-axis}

\item{dot}{The column name in \code{dt} defining the dot sizes. It must be numeric (e.g. 0.07) or character of fractions (e.g. "7/100")}

\item{dot.cex}{Positive numeric dot size scaling parameter.}

\item{qval}{The column name in \code{dt} defining the adjusted p-values for the enrichment.}

\item{term.reverse}{Logical. Controls the reverse order (y-axis) plotting.}

\item{plot.by}{Panels (facet_grids) to divide while plotting. The options are "direction" (default) and "group"}

\item{onepanel}{Logical (default FALSE). If TRUE, removes the multiple panels (no facet_grid)}

\item{arrows}{Logical (default FALSE). Should "direction" arrows be plotted inside dots?}

\item{arrow.col}{Color of the \code{arrows}. Default is "beige".}

\item{ID}{The column name in \code{dt} defining the ID of the enriched term.}

\item{Description}{The column name in \code{dt} defining the description of the enriched term.}
}
\value{
ggplot2 object for dotplot
}
\description{
Dotplot representation for enrichment results
}
\examples{
## Read sample data
dt <- system.file("extdata", "sample_gsea_kegg.tsv", package = "aamisc")
dat <- sample_read(dt)
colnames(dat)

## Dotplot with top 5 significant ontology (KEGG) terms
p <- dotplotEnrich(
        dt = dat,
        topn = 5,
        topn.pref = "dot",
        qcut = 0.05,
        nchar = 65,
        direction = "Direction",
        group = "Contrast",
        dot = "NES",
        dot.cex = 1,
        qval = "qvalue",
        term.id = "ID",
        term.name = "Description",
        term.reverse = TRUE,
        plot.by = "direction")
print(p)

## Saving as PDF file: I recommend using cairo_pdf
## DO NOT RUN
cairo_pdf("pretty_enrich_plot.pdf", height = 7, width = 7)
print(p)
dev.off()

}
