% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_script_path.R
\name{get_script_path}
\alias{get_script_path}
\title{Get current script's path}
\usage{
get_script_path(keep_name = TRUE)
}
\arguments{
\item{keep_name}{Keep name of the script. Default is TRUE.}
}
\value{
Path of the current script
}
\description{
Get current script's path
}
\references{
This script was inspired by \href{https://stackoverflow.com/a/49197577/1488825}{StackOverflow post} by user \href{https://stackoverflow.com/users/9131220/tinker}{Tinker}.
Only added a parameter called "keep_name"
}
