% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileRoot.R
\name{getFileRoot}
\alias{getFileRoot}
\title{Get file root}
\usage{
getFileRoot(x)
}
\arguments{
\item{x}{Path of a file}
}
\value{
Removed file extension. Still includes the path information
}
\description{
Get file root
}
\examples{
file <- "/home/projects/read.r"
getFileRoot(file)
#[1] "/home/projects/read"

file <- "/aaa/bbb/ccc/ddd/eee/fff/ggg.PDF"
getFileRoot(file)
#[1] "/aaa/bbb/ccc/ddd/eee/fff/ggg"

basenameLevel(getFileRoot(file))
#[1] "ggg"

basenameLevel(getFileRoot(file), dirLevel = 1)
#[1] "fff/ggg"
}
