% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMDS.R
\name{ggMDS}
\alias{ggMDS}
\title{ggplot2 wrapper for plotMDS}
\usage{
ggMDS(
  mds,
  meta,
  dim = c(1, 2),
  color.by = NULL,
  shape.by = NULL,
  text.by = NULL,
  dot.size = 2,
  text.size = 2.5,
  legend.position = "bottom"
)
}
\arguments{
\item{mds}{Output of \link{plotMDS}}

\item{meta}{Meta data of the experimental design. Column names: experimental factors to color. Row names: identical to \code{mds$cmdscale}.}

\item{dim}{Dimensions (x and y) to plot. Integer vector of length 2}

\item{color.by}{Column name of \code{meta} for colors}

\item{shape.by}{Column name of \code{meta} for shapes}

\item{text.by}{Column name of \code{meta} for texts}

\item{dot.size}{Dot size to pass into \code{geom_point(aes(size))}}

\item{text.size}{Test size to pass into \code{geom_text(aes(size))}}

\item{legend.position}{Legend position to pass into \code{theme(legend.position)}}
}
\value{
A ggplot2 object with MDS plots
}
\description{
ggplot2 wrapper for plotMDS
}
\examples{
## Meta data
meta <- data.frame(library = paste0("Lib", as.factor(1:12)),
                   patient = as.factor(rep(1:6, each = 2)),
                   disease = rep(c("healthy", "sick"), each = 6),
                   tissue = rep(c("liver","muscle"), times = 6))
meta$group <- paste(meta$disease, meta$tissue, sep = ".")
rownames(meta) <- meta$library

## Count matrix
counts <- matrix(data = round(runif(n = 1200, min = 0, max = 100)), ncol = nrow(meta))
colnames(counts) <- rownames(meta)
rownames(counts) <- paste("Gene", 1:nrow(counts), sep = "_")
head(counts)

## MDS object
mds <- limma::plotMDS(x = counts, ndim = 3, gene.selection = "pairwise", top = 500, plot = FALSE)

## Generate MDS plots with ggMDS
# Plot only points
ggMDS(mds = mds, meta = meta, dim = c(1,2))

# Add colors
ggMDS(mds = mds,
      meta = meta,
      dim = c(1,2),
      color.by = "disease")

# Add shapes
ggMDS(mds = mds,
      meta = meta,
      dim = c(1,2),
      color.by = "disease",
      shape.by = "tissue")

# Add labels
ggMDS(mds = mds,
      meta = meta,
      dim = c(1,2),
      color.by = "disease",
      shape.by = "tissue",
      text.by = "library")

# Change dimensions
ggMDS(mds = mds,
      meta = meta,
      dim = c(2,3),
      color.by = "disease",
      shape.by = "tissue",
      text.by = "library")

# Change legend location
ggMDS(mds = mds,
      meta = meta,
      dim = c(1,2),
      color.by = "disease",
      shape.by = "tissue",
      text.by = "library",
      legend.position = "right")

# Edit other ggplot2 options by creating a ggplot2 object
p <- ggMDS(mds = mds,
           meta = meta,
           dim = c(1,2),
           color.by = "disease",
           text.by = "library")
p
p + ggplot2::theme_minimal()

\dontrun{
# Make a panel of all dimensions
p1 <- ggMDS(mds = mds,
            meta = meta,
            dim = c(1,2),
            color.by = "disease",
            text.by = "library",
            legend.position = "right")
p2 <- ggMDS(mds = mds,
            meta = meta,
            dim = c(1,3),
            color.by = "disease",
            text.by = "library",
            legend.position = "right")
p3 <- ggMDS(mds = mds,
            meta = meta,
            dim = c(2,3),
            color.by = "disease",
            text.by = "library",
            legend.position = "right")
patchwork::wrap_plots(list(p1, p2, p3)) + patchwork::plot_layout(guides = 'collect')
}
}
