% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVolcano.R
\name{ggVolcano}
\alias{ggVolcano}
\title{ggplot2 wrapper for volcano plots}
\usage{
ggVolcano(
  x,
  fdr = 0.05,
  fdr.column = "FDR",
  pvalue.column = NULL,
  logFC = 0,
  logFC.column = "logFC",
  col = "gray",
  col.sig = c("blue", "red"),
  alpha = 1,
  text.column,
  text2plot = NULL,
  text.size = 3,
  dot.size = 1,
  linetype = c("blank", "solid", "dashed", "dotted", "dotdash", "longdash", "twodash")[3],
  xlab = NULL,
  ylab = NULL,
  yaxis = c("fdr", "pvalue")[1],
  dge.info = TRUE,
  dge.info.loc = 0.1
)
}
\arguments{
\item{x}{Table (\link{data.frame} or \link{data.table}) of differential expression results. See \code{limma::topTable} output as an example.}

\item{fdr}{FDR cutoff. Default = 0.05}

\item{fdr.column}{Name of the column storing FDR values}

\item{pvalue.column}{Name of the column storing (raw) p-values. Default = NULL. Required if \code{yinfo = "pvalue"}}

\item{logFC}{logFC cutoff. Default = 0}

\item{logFC.column}{Name of the column storing logFC (log2 fold change) values}

\item{col}{Color of dots}

\item{col.sig}{Color of significant dots. If 2 colors are picked, they will be used for 'down-regulated' and 'up-regulated' genes, respectively}

\item{alpha}{Transparency of dots. Range from 0 (fully tranparent) to 1 (opaque). Default = 1}

\item{text.column}{Name of the column storing text (e.g. genes) for labeling}

\item{text2plot}{Show text for dot. Character vector of \code{text} column in \code{x}. Default = NULL}

\item{text.size}{Text size to pass into \code{geom_text(aes(size))} Default = 3}

\item{dot.size}{Dot size to pass into \code{geom_point(aes(size))}. Default = 1}

\item{linetype}{Linetype of the signficance cutoffs. Default 'dashed'.}

\item{xlab}{X-axis label. Default= NULL}

\item{ylab}{Y-axis label. Default= NULL}

\item{yaxis}{Plot 'p-value' or 'FDR' on y-axis. Default 'FDR'}

\item{dge.info}{Boolean for showing differentially expressed gene (protein etc.) numbers. Default = TRUE}

\item{dge.info.loc}{The location of the \code{dge.info}. 0 is bottom and 1 is top on y-axis. Default = 0.1}
}
\value{
A ggplot2 object
}
\description{
ggplot2 wrapper for volcano plots
}
\examples{
## Generate some data
n <- 5000
set.seed(666)
dge <- data.frame(Tags = paste0("Gene_", 1:n),
                  logFC = rnorm(n, mean = 0, sd = 1),
                                    pval = runif(n))
dge[1:100, "pval"] <- runif(100, min = 0, max = 0.001)
dge <- dge[order(dge$pval),]
dge$FDR <- p.adjust(dge$pval, method = "BH")

## Create volcano plot
ggVolcano(x = dge, fdr.column = "FDR", logFC.column = "logFC", fdr = 0.05)

## Change FDR cutoff
ggVolcano(x = dge, fdr.column = "FDR", logFC.column = "logFC", fdr = 0.10)

## Change logFC cutoff
ggVolcano(x = dge, fdr.column = "FDR", logFC.column = "logFC", fdr = 0.05, logFC = 1)

## Plot raw p-values on y-axis (keeps the FDR cutoff)
ggVolcano(x = dge, fdr.column = "FDR", logFC.column = "logFC", fdr = 0.05, logFC = 1,
          pvalue.column = "pval", yaxis = "pvalue")

## Add text for genes
text2plot <- dge[1:10, "Tags"]
ggVolcano(x = dge, fdr.column = "FDR", logFC.column = "logFC", fdr = 0.05,
          text.column = "Tags", text2plot = text2plot)
ggVolcano(x = dge, fdr.column = "FDR", logFC.column = "logFC", fdr = 0.05, logFC = 1,
          text.column = "Tags", text2plot = text2plot)
}
