% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepExact.R
\name{grepExact}
\alias{grepExact}
\title{Find exact patterns using grep/grepl}
\usage{
grepExact(pattern, x, boolean = TRUE, ...)
}
\arguments{
\item{pattern}{Character vector of patterns. See \link{grep} for further details}

\item{x}{Character vector where patterns are sought. See \link{grep} for further details}

\item{boolean}{If TRUE (default) \link{grepl} will be used. If FALSE, \link{grep} will be used}

\item{...}{Further parameters to be passed on \link{grep} or \link{grepl}}
}
\value{
A vector of pattern matches. If boolean = TRUE, a boolean vector of matches will be returned. If FALSE, a numeric vector of indices with correct matches will be returned. Can be modified by using other parameters to be passed on \link{grep} or \link{grepl}. See examples below.
}
\description{
Find exact patterns using grep/grepl
}
\examples{
x <- c("a", "aa", "b", "bb", 1, 11, 111, 2, 22, 222)
y <- c("a")

# grep/grepl finds multiple matches
grepl(y, x)

# grepExact finds exact matches
grepExact(y, x, boolean = FALSE)
grepExact(y, x, boolean = TRUE)

# grep/grepl works with one string at a time
y <- c("a", 1)
grepl(y, x)
# although it can be modified with a lapply wrapper
unlist(lapply(y, function(a){grep(a,x)}))
# However, grepExact works as intended
grepExact(y, x, boolean = FALSE)
grepExact(y, x, boolean = TRUE)
}
