% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logAmp2relAmp.R
\name{logAmp2relAmp}
\alias{logAmp2relAmp}
\title{Convert log2 absolute amplitudes to relative amplitudes}
\usage{
logAmp2relAmp(x)
}
\arguments{
\item{x}{Absolute amplitude of log2 scale data. Input as numeric vector.}
}
\value{
Relative amplitude (linear scale). Output as numeric vector
}
\description{
Convert log2 absolute amplitudes to relative amplitudes
}
\details{
Below are some details about the different amplitude definitions in cosinor regression:
\itemize{
\item Absolute amplitude \code{A_abs} is the amplitude calculated in a cosinor regression
\item Log amplitude \code{A_log} is the amplitude of the log2 transformed cosinor curve
\item Relative amplitude \code{A_rel} is calculated by dividing the \code{A_abs} by the mean value
\item Fold change amplitude \code{A_fc} is fold change of relative amplitude
}

\deqn{ 2A_{log} = log_{2}(A_{fc}) }
\deqn{ 2A_{fc} = (1+A_{rel}) / (1-A_{rel}) = 2^{2(A_{log})} }, therefore
\deqn{ A_{rel} = (2^{2(A_{log})}-1)/(2^{2(A_{log})}-1) }
}
\examples{

}
