% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixScale.R
\name{matrixScale}
\alias{matrixScale}
\title{Scale a matrix using group information across the columns}
\usage{
matrixScale(x = x, id = rep("A", dim(x)[2]), center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{Input matrix}

\item{id}{A vector of group IDs. The length of this vector should be equal to the column number of input matrix.}

\item{center}{Should the data be centered? See \code{\link{scale}} for details}

\item{scale}{Should the data be scaled? See \code{\link{scale}} for details}
}
\value{
A groupwise column scaled matrix
}
\description{
Scale a matrix using group information across the columns
}
\examples{
# Create a matrix with column-wise group information
set.seed(666)
x <- matrix(rnorm(75), 15, 5)
print(x)
groups <- c(rep("A",2), rep("B",3))
colnames(x) <- uniqueNames(groups)
y <- matrixScale(x = x, id = groups, center = TRUE, scale = TRUE)
print(y)

# Group IDs do not need to be ordered
set.seed(666)
groups2 <- sample(groups)
colnames(x) <- uniqueNames(groups2)
y2 <- matrixScale(x = x, id = groups2, center = TRUE, scale = TRUE)
print(y2)
}
