% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveMeDataTable.R
\name{moveMeDataTable}
\alias{moveMeDataTable}
\title{Move columns in a \code{data.table}}
\usage{
moveMeDataTable(data, tomove, where = "last", ba = NULL)
}
\arguments{
\item{data}{A \code{data.table} object.}

\item{tomove}{The column to move.}

\item{where}{Where to move the column(s)?
\itemize{
    \item \code{last} Moves the column to the last (default)
    \item \code{first} Moves the column to the first
    \item \code{before} Moves the column before \code{ba}
    \item \code{after} Moves the column after \code{ba}
    }}

\item{ba}{The column that points "before" or "after" locations used in \code{where}}
}
\value{
Re-ordered \code{data.table}
}
\description{
This function helps moving the columns in big \code{\link{data.table}}.
}
\examples{
DT <- data.table(A=sample(3, 10, TRUE),
                B=sample(letters[1:3], 10, TRUE), C=sample(10))
DT <- moveMeDataTable(DT, "C", "after", "A")
}
\references{
\href{https://stackoverflow.com/questions/18339370/reordering-columns-in-a-large-dataframe}{StackOverflow Thread}
}
\author{
User \href{https://stackoverflow.com/users/5973989/moun}{Moun} and \href{https://stackoverflow.com/users/1270695/a5c1d2h2i1m1n2o1r2t1}{A5C1D2H2I1M1N2O1R2T1} from StackOverflow.
}
