% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qvalue.R
\name{Qvalue}
\alias{Qvalue}
\title{Calculate q-value by \code{\link{qvalue}}. If it fails, calculated adjusted p-value by \code{\link{p.adjust}}}
\usage{
Qvalue(x, alternative.method = "BH")
}
\arguments{
\item{x}{p-values}

\item{alternative.method}{If \code{\link{qvalue}} fails to calculate, function defaults to \code{\link{p.adjust}}. \code{alternative.method}
will be used as \code{method} in \code{\link{p.adjust}}. Default is "BH".}
}
\value{
q-values. If q-value cannot be calculated by \code{\link{qvalue}}, it throws an error. The \code{\link{Qvalue}} catches
this error without stopping and changes the p-value adjustment to \code{\link{p.adjust}}.
}
\description{
Calculate q-value by \code{\link{qvalue}}. If it fails, calculated adjusted p-value by \code{\link{p.adjust}}
}
\examples{
# Q-values can be calculated
p <- c(runif(n = 100, min = 0, max = 0.05),
      runif(n = 900, min = 0, max = 1))
q <- Qvalue(p)
par(mfrow = c(1,2))
hist(p); hist(q)

# Q-values cannot be calculated
p <- c(runif(n = 1, min = 0, max = 0.05),
      runif(n = 9, min = 0, max = 1))
q <- Qvalue(p)
par(mfrow = c(1,2))
hist(p); hist(q)
}
