% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad2hrs.R
\name{rad2hrs}
\alias{rad2hrs}
\title{Radian to hours}
\usage{
rad2hrs(rad)
}
\arguments{
\item{rad}{radian}
}
\value{
hours
}
\description{
Radian to hours
}
\examples{
degree <- c(0, 90, 180, 270, 360)
x <- degree*2*pi/360
rad2hrs(x)
}
