% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainAmp.R
\name{rainAmp}
\alias{rainAmp}
\title{\code{rain} wrapper with amplitude/phase estimations from \code{HarmonicRegression}}
\usage{
rainAmp(
  x,
  times,
  Period = 24,
  Method = c("independent", "longitudinal")[2],
  pAdjMethod = "BH",
  logged = FALSE,
  trend = TRUE,
  trend.pol.degree = 1,
  threads = NULL,
  ...
)
}
\arguments{
\item{x}{Numeric matrix. Rows = times, Columns = Gene names. See \link{rain} for details. Rows (times) should be sorted}

\item{times}{Numeric vector for times. Length should be equal to the row length of \code{x}}

\item{Period}{Period of \link{harmonic.regression}. Default = 24 (hours).}

\item{Method}{Can be either \code{independent} or \code{longitudinal}. See \link{rain} for details.}

\item{pAdjMethod}{P-value adjustment method. Default is \code{BH}. See \link{p.adjust} for details.}

\item{logged}{Logical. Is input matrix, \code{x}, log2 transformed?}

\item{trend}{Logical. Eliminate trend? See \link{harmonic.regression}'s \code{trend.eliminate}}

\item{trend.pol.degree}{Degree of polynomial. Numeric. If \code{trend = TRUE}, a polynomial parameter will be added to harmonic regression. See \link{harmonic.regression}'s \code{trend.degree}}

\item{threads}{Number of threads for parallel processing. Experimental at the moment.}

\item{...}{Further parameters can be passed to \link{rain}, \link{harmonic.regression}}
}
\value{
Table with the circadians results
\itemize{
\item Tags: Gene names (corresponds to the column names of input matrix \code{x})
\item pVal: p-value calculated by \link{rain}
\item phase: phase calculated by \link{rain}
\item peak.shape: peak.shape calculated by \link{rain}
\item period: period of the rhythmicity
\item pVal.adj: \link{rain} p-value adjusted for multiple testing
\item peak: peak time calculated from \link{rain}
\item trough: trough time calculated by \link{rain}
\item hr.amplitude: relative amplitude calculated by \link{harmonic.regression}
\item hr.phase: phase (peak time) calculated by \link{harmonic.regression}
\item hr.pVal: p-value calculated by \link{harmonic.regression}
\item hr.qVal: q-value calculated by \link{harmonic.regression}
}
}
\description{
This function runs \link{rain} and calculates the amplitude and phases using \link{HarmonicRegression}. Input can be log2 transformed data (\code{logged = TRUE}) and if so, the relative amplitudes will be calculated.
}
\examples{
##------------------------------------------------------------------------
## create a circadian dataset with different noise levels
##------------------------------------------------------------------------
noise.levels <- c(0, 0.02, 0.05) # range is [0,1]
timeStart <- 0
timeEnd <- 48
time <- timeStart:timeEnd
amplitude <- 0.4 # this is relative amplitude
baseline <- 1
period <- 24

##-----------------------------------------------
## Linear scale
##-----------------------------------------------
test_linear <- apply(matrix(noise.levels, nrow = 1), 2, function(noise){
  timecourse = baseline + amplitude * cos((time) / period * 2 * pi) + rnorm(timeEnd, 0, noise)
})
colnames(test_linear) <- paste("Gene", 1:ncol(test_linear), sep = "_")
rownames(test_linear) <- paste("Time", 1:nrow(test_linear), sep = "_")

## Plot the time series (linear scale)
plot(time, test_linear[,1], xlab = "Time", ylab = "Linear Scale",
     sub = paste("Relative amplitude is", amplitude))

res_linear <- rainAmp(x = test_linear,
                      times = time,
                      Period = period,
                      Method = "independent",
                      pAdjMethod = "BH",
                      logged = FALSE, # Note that this is linear scale
                      trend = FALSE)
res_linear[1,]
## Note that amplitude estimation (hr.amplitude) is ~0.4, which is what we set)
##     Tags         pVal phase peak.shape period     pVal.adj peak trough hr.amplitude hr.phase      hr.pVal      hr.qVal
## 1: Gene_1 1.022744e-45     1         12     24 3.068233e-45    2     14    0.3967611 24.00000 0.000000e+00 0.000000e+00

##-----------------------------------------------
## Log2 cale
##-----------------------------------------------
test_log2 <- log2(test_linear)

## Plot the time series (log2 scale)
plot(time, test_log2[,1], xlab = "Time", ylab = "Log2 Scale",
     sub = paste("Relative amplitude is", amplitude))

res_log2 <- rainAmp(x = test_log2,
                    times = time,
                    Period = period,
                    Method = "independent",
                    pAdjMethod = "BH",
                    logged = TRUE, # Note that this is log2 scale
                    trend = FALSE)
res_log2[1,]
## Note that amplitude estimation (hr.amplitude) is ~0.4, which is what we set)
##      Tags         pVal phase peak.shape period     pVal.adj peak trough hr.amplitude hr.phase      hr.pVal      hr.qVal
## 1: Gene_1 1.022744e-45     1         12     24 3.068233e-45    2     14    0.3934892       24 8.006058e-46 2.401817e-45

}
