% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_read.R
\name{sample_read}
\alias{sample_read}
\title{Read sample datasets in \link{aamisc}}
\usage{
sample_read(file)
}
\arguments{
\item{file}{Full path of the sample file}
}
\value{
\code{data.table}
}
\description{
Read sample datasets in \link{aamisc}
}
\examples{
dt <- system.file("extdata", "sample_gsea_kegg.tsv", package = "aamisc")
sample_read(dt)

}
