% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueNames.R
\name{uniqueNames}
\alias{uniqueNames}
\title{Generate unique names of a vector with repeated elements}
\usage{
uniqueNames(x, sep = "_")
}
\arguments{
\item{x}{Input vector.}

\item{sep}{A character string to separate the terms. Default is "_".}
}
\value{
Vector of unique names created by
}
\description{
Generate unique names of a vector with repeated elements
}
\examples{
x <- letters[c(1,2,2,3,3,3,4,6)]
res <- uniqueNames(x, sep = "_")
print(res)
res2 <- uniqueNames(x, sep = ".")
print(res2)
}
