% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplotEnrich.R
\name{dotplotEnrich}
\alias{dotplotEnrich}
\title{Generate unique names of a vector with repeated elements}
\usage{
dotplotEnrich(
  dt,
  topn = 10,
  topn.pref = c("q", "dot"),
  qcut = 0.05,
  nchar = 60,
  direction = "DEstatus",
  group = "Condition",
  dot = "GeneRatio",
  dot.cex = 1,
  q = "p.adjust",
  term.id = "ID",
  term.name = "Description",
  term.reverse = FALSE,
  plot.by = c("direction", "group")[1],
  pdf.name = NULL,
  return.plotObj = FALSE
)
}
\arguments{
\item{dt}{data.table or data.frame with the enrichment results. See details for necessary information.}

\item{topn}{Top \code{n} enriched terms to be plotted}

\item{topn.pref}{The ranking method to identify \code{topn} terms. Can be either \code{q} or \code{dot}.}

\item{qcut}{The adjusted p-value cutoff for enriched terms.}

\item{nchar}{Maximum characters to be shown on on y-axis. If the term contains more characters than \code{nchar}, it will be abbreviated with "(...)"}

\item{direction}{The column name in \code{dt} defining the direction  (e.g. up, down, all) of the enrichment}

\item{group}{The column name in \code{dt} defining the groups for each enrichment. The contents will be written on x-axis}

\item{dot}{The column name in \code{dt} defining the dot sizes. It must be numeric (e.g. 0.07) or character of fractions (e.g. "7/100")}

\item{dot.cex}{Positive numeric dot size scaling parameter.}

\item{q}{The column name (numeric) in \code{dt} defining the adjusted p-values for the enrichment.}

\item{term.reverse}{Logical. Controls the reverse order (y-axis) plotting.}

\item{plot.by}{Panels (facet_grids) to divide while plotting. The options are "direction" (default) and "group"}

\item{pdf.name}{Generates a pdf file if not \code{NULL}}

\item{return.plotObj}{Boolean controlling whether a \code{ggplot2} object should be returned. Default is FALSE.}

\item{ID}{The column name (numeric) in \code{dt} defining the ID of the enriched term.}

\item{Description}{The column name (numeric) in \code{dt} defining the description of the enriched term.}
}
\value{
dotplot
}
\description{
Generate unique names of a vector with repeated elements
}
